+++ examples/weaviate/vector_search.py
@@ -1,5 +1,5 @@
 from neo4j import GraphDatabase
-from neo4j_genai.retrievers.external.weaviate import WeaviateNeo4jRetriever
+from neo4j_genai.retrievers import WeaviateNeo4jRetriever
 from weaviate.connect.helpers import connect_to_local
 
 from examples.embedding_biology import EMBEDDING_BIOLOGY
