+++ src/neo4j_genai/__init__.py
@@ -12,16 +12,3 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
-from .retrievers.hybrid import HybridCypherRetriever, HybridRetriever
-from .retrievers.text2cypher import Text2CypherRetriever
-from .retrievers.vector import VectorCypherRetriever, VectorRetriever
-
-
-__all__ = [
-    "VectorRetriever",
-    "VectorCypherRetriever",
-    "HybridRetriever",
-    "HybridCypherRetriever",
-    "Text2CypherRetriever",
-]
