+++ src/neo4j_genai/retrievers/external/pinecone/__init__.py
@@ -12,7 +12,3 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
-from .pinecone import PineconeNeo4jRetriever
-
-__all__ = ["PineconeNeo4jRetriever"]
