+++ src/neo4j_genai/retrievers/external/weaviate/__init__.py
@@ -12,7 +12,3 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-
-from .weaviate import WeaviateNeo4jRetriever
-
-__all__ = ["WeaviateNeo4jRetriever"]
