+++ src/neo4j_genai/retrievers/external/weaviate/types.py
@@ -25,7 +25,7 @@ from pydantic import (
 from weaviate.client import WeaviateClient
 from weaviate.collections.classes.filters import _Filters
 
-from neo4j_genai.retrievers.utils import validate_search_query_input
+from neo4j_genai.utils import validate_search_query_input
 from neo4j_genai.types import Neo4jDriverModel, EmbedderModel
