+++ tests/e2e/pinecone_e2e/test_pinecone_e2e.py
@@ -19,8 +19,8 @@ from unittest.mock import MagicMock
 
 import pytest
 from langchain_huggingface.embeddings import HuggingFaceEmbeddings
+from neo4j_genai.retrievers import PineconeNeo4jRetriever
 from neo4j_genai.embedder import Embedder
-from neo4j_genai.retrievers.external.pinecone import PineconeNeo4jRetriever
 from neo4j_genai.types import RetrieverResult, RetrieverResultItem
 from pinecone import Pinecone
