+++ tests/e2e/test_text2cypher_e2e.py
@@ -1,5 +1,5 @@
 import pytest
-from neo4j_genai import Text2CypherRetriever
+from neo4j_genai.retrievers import Text2CypherRetriever
 from unittest.mock import MagicMock
 from neo4j_genai.types import RetrieverResult, RetrieverResultItem
