+++ tests/unit/retrievers/external/test_weaviate.py
@@ -19,7 +19,7 @@ from unittest.mock import MagicMock
 
 import neo4j
 from neo4j_genai.retrievers.external.utils import get_match_query
-from neo4j_genai.retrievers.external.weaviate import WeaviateNeo4jRetriever
+from neo4j_genai.retrievers import WeaviateNeo4jRetriever
 from neo4j_genai.types import RetrieverResult, RetrieverResultItem
 from weaviate.client import WeaviateClient
